Covered a lot of loose holes in terms of cleanup.
Under no circumstances should the following happen:
 - local socket remains after vs-X exits
 - real X server keeps running after vs-X exits

If these happen, then X server fail to start 
later.

The current changes are an effort to fix these.
I have moved the signal handlers to an earlier 
position, ensured that vs-X kills the X server and
waits for it before it exits, and also ensured
that the same things happen when the SSM exits
midway. Seems to be working with whatever testing
I have done already.