Changed the behaviour of this script.
With the changes, the VirtualGL client program is started
if the --x11-draw is not specified on the command line.

The user can point the script to a VirtualGL client program
using the environment variables VGL_CLIENT and VGL_PORT. These
options could be used when the final display is not on the
node from where this script is run.