Partial solution for ticket "Prioritize GPU allocation" https://sourceforge.net/apps/trac/vizstack/ticket/14
vsapi.py adds a getAllocationWeight() to every
VizResource.

metascheduler.py uses these weights to sort
available resources. The explicit index based
sorting for X servers has been removed.

The only unimplemented item(s) in this changeset is
  - GPUs that are part of Tiled Displays were
    supposed to be prioritized lower. This hasn't
    been done yet

Also, connected display devices are not contributing to
the weight. They do in the code, but not in effect.
This is the case since this information is not generated 
as part of the configuration. Once we add code to the
configuration process to retain connected display device
information, then this will automatically take effect.