python/vsapi.py
Implemented an allocation 'bias' value for 
each VizResource.

This influences the getAllocationWeight() function,
thus automatically altering the allocation order
for resources.

sbin/vs-ssm

The bias value is used to ensure that GPUs that
are part of tiled displays are prioritised to
be allocated last. To achieve this, the SSM
finds out all GPUs that are part of tiled displays,
and assigns a bias of 100000 to all fixed GPUs
that are part of these.