Added primitive support for bezels. These DO NOT contribute to tiled displays yet !
Each display device can now have
 - a physical dimension (width by height), 
   specified in millimeters
 - four bezels - left, right, bottom, top,
   also specified in millimeters

This information is used to compute the
size of the bezel (in pixels) for each mode.

It is also possible to override the individual
bezel sizes per resolution; to do this, the
bezel information can be included per mode.
The specified value will override any value
computed by using the physical dimensions.