Fixed stereo usage. Stereo cannot now be specified on the command line. If the tiled display has stereo configured, then it will run with the stereo setting.
One reason for this is to keep the script immune to
added stereo modes. Besides a stereo tiled display will
mostly not be suitable for another kind of stereo display,
so it makes sense from a user POV to keep it that way.

In practise, stereo display surfaces seem to be running
stereo most of the time (especially CRT projector based,
since the size of the image is related to the modeline).

In case the users need a non-stereo tiled display that is
exactly the same as a stereo display but with the stereo
disabled, they can always get one created.

Doing this programmatically is very easy with the vsapi;
and the power users could do this if they wish.