Removed the restriction that a GPU is limited to max display port number of 3. This restriction came from the Quadro FX 5800 series cards we were using.
The limit was broken by the Quadro NVS 420, which
has 2 GPUs, but each of them can drive either
DFP-0, DFP-1 or DFP-2, DFP-3.

Simon Fowler encountered this as a problem, so
I fixed it.

I have also added a template for the Quadro NVS 420
card, since I don't expect vs-generate-xconfig
to be able to handle the intricacies of the NVS
420 !