Fix for ticket 36.
https://sourceforge.net/apps/trac/vizstack/ticket/36

The fix ensures that leading zeroes are used in
nodename expansion when appropriate. Node name
expansion was being used in two places :
  - SLURM scheduler classes
  - configure script.

So, I centralized this into vsutil. That caused
other pythonic probles like circular references
in modules. I had to fix that by
  - removing all instances of "from module import symbol"
  - stopping usage of config file paths in default args.

A look at the diffs can be confusing due to these
changes, but the principal changes are in
  - vsutil.py
  - slurmscheduler.py
  - vs-configure-system