Fix for SF bug : 3101341
When I implemented the weight mechanism, I used the
word "allocationBias" in the XML. This was later
changed to "weight" to make the XML easy to understand.
Unfortunately, the change wasn't propagated to vsutil,
so the node weights were all 0, and getting ignored.

Fixed code to use the tag "weight" rather than "allocationBias".